+++ CHANGES.md
@@ -42,6 +42,7 @@ Change Log
 * Fix bug `TableColumn.type = scalar` even if there were no values.
 * `TableColumn.type = enum` requires at least 2 unique values (including null) to be selected by default
 * `TableStyle` `startDates`, `finishDates` and `timeIntervals` will only set values for valid `rowGroups` (invalid rows will be set to `null`). For example, this means that rows with invalid regions will be ignored.
+* Add "Disable style" option to `TableMixin.styleDimensions` - it can be enabled with `TableTraits.showDisableStyleOption`
 * [The next improvement]
 
 #### 8.1.4